blame-tree: pass repository pointer into `blame_tree_init()`
Teach `cmd_blame_tree()` to pass `the_repository` into `blame_tree_init()`.

This makes the merge-based integration of v2.41.0 match the
candidate rebase-based integration and cuts down on noise
in the diffs.

I'm not sure of the full history of this change. It has never
appeared in the tree before when using the merge-based history,
but was (rebased-and-squashed? into) the initial circa-2012 code
in the rebase-based history, so it looks like it was always this
way.

Either way, it doesn't hurt anything thing to make them consistent
and reduce future diffs as we integrate and sync future batches.

NEEDSWORK: The code in `blame-tree.c` has a couple of references
to `the_repository`.  These should be investigated.  One has a
`struct blame_tree *` and could be easily fixed.  One does not.