config: introduce git_config_float()
Future commits will want to parse a floating point value from
configuration, but we have no way to parse such a value prior to this
patch.

The core of the routine is implemented in git_parse_float(). Unlike
git_parse_unsigned() and git_parse_signed(), however, the function
implemented here only works on type "float", and not related types like
"double", or "long double".

This is because "double" and "long double" use different functions to
convert from ASCII strings to floating point values (strtod() and
strtold(), respectively). Likewise, there is no pointer type that can
assign to any of these values (except for "void *"), so the only way to
define this trio of functions would be with a macro expansion that is
parameterized over the floating point type and conversion function.

That is all doable, but likely to be overkill given our current needs,
which is only to parse floats.

Signed-off-by: Taylor Blau <me@ttaylorr.com>