blame-tree: avoid pretending to be pack metadata
When we merged Git v2.36, it introduced fine-grained `fsync()` support
via 844a8ad4f86 (core.fsync: add configuration parsing, 2022-03-10), we
needed to adjust the `blame-tree` code to pick an `fsync()` category to
allow for that fine granularity.

At the time, the decision was made to simply piggy-back onto
`FSYNC_COMPONENT_PACK_METADATA` and be done with it.

However, the notion is quite conceivable that pack metadata may be a lot
more precious than blame-tree data.

That fact, combined with the fact that 83937e95928 (Merge branch
'ns/batch-fsync', 2022-06-03) introduced many more `fsync()` categories,
indicates strongly that we want a _separate_ category for `blame-tree`.

Make it so.