pack-bitmap.c: extract a "struct bitmap_walk"
Our `struct bitmap_index` has two internal bitmap fields corresponding
to the interesting and uninteresting sides of its most recent
reachability traversal. These are known as "result" and "haves",
respectively.

In a forthcoming change, the concept of a bitmap will be extended to
support incremental multi-pack indexes, with each layer of the
incremental MIDX maintaining its own bitmap. Each bitmap's objects
appear in the same order, with new objects being appended onto the end
of the existing object order.

While each bitmap layer could maintain its own view of the object walk
(setting only the relevant bits in its own "result" and "haves"
bitmaps), it is far more straightforward to conduct the whole walk using
only a single pair of these bitmaps at the end of the MIDX bitmap chain.

But it doesn't suffice to merely recurse down on the struct bitmap_index
itself, since doing so would cause us to set bits in each layer's
results/haves bitmap. Instead, following commits will pass a pointer to
the tip bitmap's bitmap_walk down when making recursive calls so that
only that pair of bitmaps have their bits set when performing a walk.

For now this is a non-functional refactoring in preparation for the
subsequent commits.

Signed-off-by: Taylor Blau <me@ttaylorr.com>