pack-bitmap.c: use tip bitmap in traversal functions
The pack-bitmap machinery uses traverse_bitmap_commit_list() to present
the list of reachable objects for some query to the caller by invoking
the callback function on its set bits, respecting the type-level filters
on the reachability query itself.

These functions work by:

  - For each object type included in the reachability query, enumerate
    the set of reachable objects using the corresponding type bitmap as
    a filter.

  - Performing a similar function using show_extended_objects(), but
    checking instead the actual obj->type (since we don't maintain
    separate type-level bitmaps for the eindex, as we have actual object
    structs in that case).

Prepare these functions for incremental MIDX bitmaps by teaching them to
enumerate the results for objects from each layer's type bitmaps. This
step works recursively, since each layer only stores type-level
information for objects that are contained in the corresponding MIDX
layer.

(Alternatively, we could OR together the type-level bitmaps and avoid
the recursion, but I find the result below easier to read and reason
about).

The transformation here is mostly straightforward as it involves
recursively calling traverse_bitmap_commit_list_1() for each bitmap
layer. Note, however, that the recursive call *only* uses the
bitmap_walk structure pertaining to the top-most layer, since that is
where the actual haves/wants bitmaps are stored for the entire traversal.

Note further that the actual traversal function itself does not yet
understand how to perform a multi-bitmap traversal. This will be
implemented in a future commit.

Signed-off-by: Taylor Blau <me@ttaylorr.com>