object-store: pass down repository to `each_packed_object_fn`
The `each_packed_object_fn` defines the type of the function called for
each packed object. In some of the implementations, we require the
repository state. So let's modify the function type to also include a
repository object and modify all implementations to receive the object
and utilize that instead of global state.

Signed-off-by: Karthik Nayak <karthik.188@gmail.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>